# api/openapi.yaml
openapi: 3.0.0
info:
  title: API de servidores sancionados
  version: 1.1.0
externalDocs:
  description: Más información
  url: https://github.com/PDNMX/api_docs
servers:
- url: /
paths:
  /v1/ssancionados:
    post:
      tags:
      - ssancionados
      summary: Muestra información de los servidores públicos sancionados permitiendo
        busquedas avanzadas.
      operationId: post_ssancionados
      requestBody:
        description: JSON para peticiones de búsqueda avanzada
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/post_query'
        required: true
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ssancionados'
        "400":
          description: Página inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resError'
        "401":
          description: No tienes permisos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resError'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resError'
      security:
      - ssancionados_auth: []
      x-swagger-router-controller: Ssancionados
  /v1/ssancionados/dependencias:
    get:
      tags:
      - ssancionados
      summary: Dependencias donde se cometió la sanción
      operationId: get_dependencias
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dependencias'
        "400":
          description: Página inválida
        "403":
          description: No tienes permisos
      security:
      - ssancionados_auth: []
      x-swagger-router-controller: Ssancionados
components:
  schemas:
    ssancionados:
      required:
      - pagination
      - results
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/ssancionados_pagination'
        results:
          type: array
          items:
            $ref: '#/components/schemas/ssancionados_results'
      example:
        pagination:
          hasNextPage: true
          pageSize: 10
          page: 10
          totalRows: 100
        results:
        - autoridadSancionadora: Organo Interno de Control
          institucionDependencia:
            clave: TRE345
            siglas: SHCP
            nombre: Secretaría de Hacienda y Crédito Público
          tipoFalta:
            descripcion: Administrativa grave
            clave: NAD
            valor: ADMINISTRATIVA GRAVE
          resolucion:
            fechaResolucion: 2018-05-21T00:00:00.000+0000
            url: https://ejemplo.mx/archivo/documento.pdf
          expediente: ABC123
          multa:
            monto: 30000.19
            moneda:
              clave: MXN
              valor: Peso Mexicano
          tipoSancion:
          - clave: O
            valor: AMONESTACIÓN
          - clave: M
            valor: MULTADO
          documentos:
          - descripcion: Descripción del documento
            fecha: 2019-11-20T00:00:00.000+0000
            tipo: CONSTANCIA_SANCION
            titulo: Constancia de sanción...
            id: doc-1
            url: https://ejemplo.mx/archivo/documento.pdf
          - descripcion: Descripción del documento
            fecha: 2019-11-20T00:00:00.000+0000
            tipo: CONSTANCIA_SANCION
            titulo: Constancia de sanción...
            id: doc-1
            url: https://ejemplo.mx/archivo/documento.pdf
          fechaCaptura: 2017-07-21T17:32:28Z
          causaMotivoHechos: Abuso de autoridad
          observaciones: ABC
          servidorPublicoSancionado:
            puesto: Director de área
            primerApellido: Pérez
            genero:
              clave: M
              valor: MASCULINO
            segundoApellido: López
            nivel: ABCD1234
            rfc: PAT21343213
            curp: ABCD780814MDFAXL
            nombres: Carlos
          id: "1"
          inhabilitacion:
            fechaInicial: 2018-08-17T00:00:00.000+0000
            plazo: 3 meses
            fechaFinal: 2018-11-17T00:00:00.000+0000
        - autoridadSancionadora: Organo Interno de Control
          institucionDependencia:
            clave: TRE345
            siglas: SHCP
            nombre: Secretaría de Hacienda y Crédito Público
          tipoFalta:
            descripcion: Administrativa grave
            clave: NAD
            valor: ADMINISTRATIVA GRAVE
          resolucion:
            fechaResolucion: 2018-05-21T00:00:00.000+0000
            url: https://ejemplo.mx/archivo/documento.pdf
          expediente: ABC123
          multa:
            monto: 30000.19
            moneda:
              clave: MXN
              valor: Peso Mexicano
          tipoSancion:
          - clave: O
            valor: AMONESTACIÓN
          - clave: M
            valor: MULTADO
          documentos:
          - descripcion: Descripción del documento
            fecha: 2019-11-20T00:00:00.000+0000
            tipo: CONSTANCIA_SANCION
            titulo: Constancia de sanción...
            id: doc-1
            url: https://ejemplo.mx/archivo/documento.pdf
          - descripcion: Descripción del documento
            fecha: 2019-11-20T00:00:00.000+0000
            tipo: CONSTANCIA_SANCION
            titulo: Constancia de sanción...
            id: doc-1
            url: https://ejemplo.mx/archivo/documento.pdf
          fechaCaptura: 2017-07-21T17:32:28Z
          causaMotivoHechos: Abuso de autoridad
          observaciones: ABC
          servidorPublicoSancionado:
            puesto: Director de área
            primerApellido: Pérez
            genero:
              clave: M
              valor: MASCULINO
            segundoApellido: López
            nivel: ABCD1234
            rfc: PAT21343213
            curp: ABCD780814MDFAXL
            nombres: Carlos
          id: "1"
          inhabilitacion:
            fechaInicial: 2018-08-17T00:00:00.000+0000
            plazo: 3 meses
            fechaFinal: 2018-11-17T00:00:00.000+0000
    post_query:
      type: object
      properties:
        pageSize:
          type: integer
          description: Número de registros a mostrar por página, el máximo permitido
            es 200 y el minimo 1, en caso de no recibir el parámetro, su valor será
            el indicado en default.
          example: 10
          default: 10
        page:
          type: integer
          description: Número de página a mostrar, en caso de no recibir el parámetro,
            su valor será el indicado en default.
          example: 10
          default: 1
        sort:
          $ref: '#/components/schemas/post_query_sort'
        query:
          $ref: '#/components/schemas/post_query_query'
    dependencias:
      type: array
      description: Instituciones o dependencias donde se cometió la sanción
      items:
        $ref: '#/components/schemas/dependencias_inner'
    resError:
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: Código alfa numérico del error de la consulta.
          example: abc123
        message:
          type: string
          description: Texto descriptivo para el usuario final del error en la consulta
          example: Error al formar la consulta, revise el formato de los campos.
      description: Estos valores son definidos por el sistema del API.
    tipoSancion:
      type: array
      description: La sanción impuesta al servidor público, puede ser más de una.
        Ej. Inhabilitado y Multado. En caso de ser OTRO, en el atributo valor se deberá
        enviar la cadena correspondiente.
      example:
      - clave: O
        valor: AMONESTACIÓN
      - clave: M
        valor: MULTADO
      items:
        $ref: '#/components/schemas/tipoSancion_inner'
    tipoFalta:
      required:
      - clave
      - valor
      type: object
      properties:
        clave:
          type: string
          description: "Clave del tipo de falta cometida. Valores permitidos\n* `NAD`\
            \ - NEGLIGENCIA ADMINISTRATIVA\n* `VPC` - VIOLACION PROCEDIMIENTOS DE\
            \ CONTRATACION\n* `VLNP` - VIOLACION LEYES Y NORMATIVIDAD PRESUPUESTAL\n\
            * `AUT` - ABUSO DE AUTORIDAD\n* `CEX` - COHECHO O EXTORSION\n* `IDSP`\
            \ - INCUMPLIMIENTO EN DECLARACION DE SITUACION PATRIMONIAL\n* `DCSP` -\
            \ DELITO COMETIDO POR SERVIDORES PUBLICOS\n* `EIFM` - EJERCICIO INDEBIDO\
            \ DE SUS FUNCIONES EN MATERIA MIGRATORIA\n* `VDH` - VIOLACIÓN A LOS DERECHOS\
            \ HUMANOS\n* `AG` - ADMINISTRATIVA GRAVE\n* `ANG` - ADMINISTRATIVA NO\
            \ GRAVE\n* `AC` - ACTO DE CORRUPCIÓN\n* `HSEX` - COMETER O TOLERAR CONDUCTAS\
            \ DE HOSTIGAMIENTO SEXUAL\n* `ASEX` - COMETER O TOLERAR CONDUCTAS DE ACOSO\
            \ SEXUAL\n* `PEC` - PECULADO\n* `DRP` - DESVÍO DE RECURSOS PÚBLICOS\n\
            * `UII` - UTILIZACIÓN INDEBIDA DE INFORMACIÓN\n* `AFN` - ABUSO DE FUNCIONES\n\
            * `ABCI` - ACTUACIÓN BAJO CONFLICTO DE INTERÉS\n* `CIND` - CONTRATACIÓ\
            N INDEBIDA\n* `EOCI` - ENRIQUECIMIENTO OCULTO U OCULTAMIENTO DE CONFLICTO\
            \ DE INTERÉS\n* `TINF` - TRÁFICO DE INFLUENCIAS\n* `ENCB` - ENCUBRIMIENTO\n\
            * `DSCT` - DESACATO\n* `OJUST` -\tOBSTRUCCIÓN DE LA JUSTICIA\n* `OTRO`\
            \ - OTRO  \nexample: AG\n"
          enum:
          - NAD
          - VPC
          - VLNP
          - AUT
          - CEX
          - IDSP
          - DCSP
          - EIFM
          - VDH
          - AG
          - ANG
          - AC
          - HSEX
          - ASEX
          - PEC
          - DRP
          - UII
          - AFN
          - ABCI
          - CIND
          - EOCI
          - TINF
          - ENCB
          - DSCT
          - OJUST
          - OTRO
        valor:
          type: string
          description: Valor del tipo de falta
          example: ADMINISTRATIVA GRAVE
        descripcion:
          type: string
          description: Descripción o nota aclaratoria.
          example: Administrativa grave
      description: Falta cometida por el servidor público
      example:
        descripcion: Administrativa grave
        clave: NAD
        valor: ADMINISTRATIVA GRAVE
    ssancionados_pagination:
      required:
      - hasNextPage
      - page
      - pageSize
      - totalRows
      type: object
      properties:
        pageSize:
          type: integer
          description: Número de registros a mostrar por página, el máximo permitido
            es 200 y el minimo 1
          example: 10
        page:
          type: integer
          description: Número de página a mostrar
          example: 10
        totalRows:
          minimum: 1
          type: number
          description: Total de registros
          example: 100
        hasNextPage:
          minimum: 1
          type: boolean
          description: indica si existe una siguiente pagina
          example: true
      example:
        hasNextPage: true
        pageSize: 10
        page: 10
        totalRows: 100
    ssancionados_institucionDependencia:
      required:
      - nombre
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la Institución o Dependencia a la que pertenece el
            Servidor Público Sancionado
          example: Secretaría de Hacienda y Crédito Público
        siglas:
          type: string
          description: Siglas de la Institución o Dependencia a la que pertenece el
            Servidor Público Sancionado
          example: SHCP
        clave:
          type: string
          description: Clave de la Institución o Dependencia a la que pertenece el
            Servidor Público Sancionado
          example: TRE345
      example:
        clave: TRE345
        siglas: SHCP
        nombre: Secretaría de Hacienda y Crédito Público
    ssancionados_servidorPublicoSancionado_genero:
      type: object
      properties:
        clave:
          type: string
          description: "Genero del Servidor Público.\n  * `F` - FEMENINO \n  * `M`\
            \ - MASCULINO\n  * `O` - OTRO\n"
          example: M
          enum:
          - F
          - M
          - O
        valor:
          type: string
          description: Valor del genero
          example: MASCULINO
      example:
        clave: M
        valor: MASCULINO
    ssancionados_servidorPublicoSancionado:
      required:
      - nombres
      - primerApellido
      - puesto
      type: object
      properties:
        rfc:
          type: string
          description: RFC del Servidor Público con homoclave
          example: PAT21343213
        curp:
          type: string
          description: CURP del Servidor Público
          example: ABCD780814MDFAXL
        nombres:
          type: string
          description: Nombre(s) del Servidor Público Sancionado
          example: Carlos
        primerApellido:
          type: string
          description: Primer apellido del Servidor Público Sancionado
          example: Pérez
        segundoApellido:
          type: string
          description: Segundo apellido del Servidor Público Sancionado
          example: López
        genero:
          $ref: '#/components/schemas/ssancionados_servidorPublicoSancionado_genero'
        puesto:
          type: string
          description: Puesto del Servidor Público Sancionado
          example: Director de área
        nivel:
          type: string
          description: Clave del nivel del puesto del Servidor Público Sancionado
          example: ABCD1234
      example:
        puesto: Director de área
        primerApellido: Pérez
        genero:
          clave: M
          valor: MASCULINO
        segundoApellido: López
        nivel: ABCD1234
        rfc: PAT21343213
        curp: ABCD780814MDFAXL
        nombres: Carlos
    ssancionados_resolucion:
      required:
      - fechaResolucion
      type: object
      properties:
        url:
          type: string
          description: URL que apunta al documento en formato digital de la resolución
            emitida por el correspondiente Organo Interno de Control
          example: https://ejemplo.mx/archivo/documento.pdf
        fechaResolucion:
          type: string
          description: Fecha en la que se emite la resolución sancionatoria. Fecha
            en formato ISO 8601
          format: date
          example: 2018-05-21
      example:
        fechaResolucion: 2018-05-21T00:00:00.000+0000
        url: https://ejemplo.mx/archivo/documento.pdf
    ssancionados_multa_moneda:
      type: object
      properties:
        clave:
          type: string
          description: ISO 4217
          example: MXN
        valor:
          type: string
          example: Peso Mexicano
      description: Tipo de moneda. Señalar tipo de moneda.
      example:
        clave: MXN
        valor: Peso Mexicano
    ssancionados_multa:
      required:
      - moneda
      - monto
      type: object
      properties:
        monto:
          type: number
          description: Monto de la multa expresado en la moneda origen
          format: float
          example: 30000.19
        moneda:
          $ref: '#/components/schemas/ssancionados_multa_moneda'
      example:
        monto: 30000.19
        moneda:
          clave: MXN
          valor: Peso Mexicano
    ssancionados_inhabilitacion:
      type: object
      properties:
        plazo:
          type: string
          description: Plazo de la inhabilitación. Ej. 2 años
          example: 3 meses
        fechaInicial:
          type: string
          description: Fecha inicial de la inhabilitación. Fecha en formato ISO 8601
          format: date
          example: 2018-08-17
        fechaFinal:
          type: string
          description: Fecha final de la inhabilitación. Fecha en formato ISO 8601
          format: date
          example: 2018-11-17
      example:
        fechaInicial: 2018-08-17T00:00:00.000+0000
        plazo: 3 meses
        fechaFinal: 2018-11-17T00:00:00.000+0000
    ssancionados_documentos:
      required:
      - descripcion
      - fecha
      - id
      - titulo
      - url
      type: object
      properties:
        id:
          type: string
          description: Puede ser el folio del documento o al menos un consecutivo
            para identificar a los documentos en el registro.
          example: doc-1
        tipo:
          type: string
          description: Tipo de documento, Valores permitidos ["RESOLUCION", "CONSTANCIA_SANCION",
            "CONSTANCIA_INHABILITACION", "CONSTANCIA_ABSTENCION"]
          example: CONSTANCIA_SANCION
          enum:
          - RESOLUCION
          - CONSTANCIA_SANCION
          - CONSTANCIA_INHABILITACION
          - CONSTANCIA_ABSTENCION
        titulo:
          type: string
          description: Título del documento
          example: Constancia de sanción...
        descripcion:
          type: string
          description: Descripción del documento
          example: Descripción del documento
        url:
          type: string
          description: Hipervínculo al documento
          example: https://ejemplo.mx/archivo/documento.pdf
        fecha:
          type: string
          description: Fecha del documento. Fecha en formato ISO 8601
          format: date
          example: 2019-11-20
      example:
        descripcion: Descripción del documento
        fecha: 2019-11-20T00:00:00.000+0000
        tipo: CONSTANCIA_SANCION
        titulo: Constancia de sanción...
        id: doc-1
        url: https://ejemplo.mx/archivo/documento.pdf
    ssancionados_results:
      required:
      - causaMotivoHechos
      - fechaCaptura
      - id
      - institucionDependencia
      - servidorPublicoSancionado
      - tipoFalta
      - tipoSancion
      type: object
      properties:
        id:
          type: string
          description: Dato que permita identificar de manera  unívoca al registro
          example: "1"
        fechaCaptura:
          type: string
          description: La fecha y hora en la que se agrega o modifica el registro.
            Fecha en formato ISO 8601
          format: date-time
          example: 2017-07-21T17:32:28Z
        expediente:
          type: string
          description: Número de Expediente del Procedimiento Administrativo de Responsabilidades
          example: ABC123
        institucionDependencia:
          $ref: '#/components/schemas/ssancionados_institucionDependencia'
        servidorPublicoSancionado:
          $ref: '#/components/schemas/ssancionados_servidorPublicoSancionado'
        autoridadSancionadora:
          type: string
          description: 'Autoridad encarcada del proceso de sanción '
          example: Organo Interno de Control
        tipoFalta:
          $ref: '#/components/schemas/tipoFalta'
        tipoSancion:
          $ref: '#/components/schemas/tipoSancion'
        causaMotivoHechos:
          type: string
          description: 'Causa o motivo de la sanción, ejemplo: Abuso de autoridad'
          example: Abuso de autoridad
        resolucion:
          $ref: '#/components/schemas/ssancionados_resolucion'
        multa:
          $ref: '#/components/schemas/ssancionados_multa'
        inhabilitacion:
          $ref: '#/components/schemas/ssancionados_inhabilitacion'
        observaciones:
          type: string
          description: 'Cualquier observación pertinente '
          example: ABC
        documentos:
          type: array
          description: Es una sección que permite adjuntar referencias a cualquier
            documento que se considere de valor para transparentar el proceso de sanción.
          items:
            $ref: '#/components/schemas/ssancionados_documentos'
      example:
        autoridadSancionadora: Organo Interno de Control
        institucionDependencia:
          clave: TRE345
          siglas: SHCP
          nombre: Secretaría de Hacienda y Crédito Público
        tipoFalta:
          descripcion: Administrativa grave
          clave: NAD
          valor: ADMINISTRATIVA GRAVE
        resolucion:
          fechaResolucion: 2018-05-21T00:00:00.000+0000
          url: https://ejemplo.mx/archivo/documento.pdf
        expediente: ABC123
        multa:
          monto: 30000.19
          moneda:
            clave: MXN
            valor: Peso Mexicano
        tipoSancion:
        - clave: O
          valor: AMONESTACIÓN
        - clave: M
          valor: MULTADO
        documentos:
        - descripcion: Descripción del documento
          fecha: 2019-11-20T00:00:00.000+0000
          tipo: CONSTANCIA_SANCION
          titulo: Constancia de sanción...
          id: doc-1
          url: https://ejemplo.mx/archivo/documento.pdf
        - descripcion: Descripción del documento
          fecha: 2019-11-20T00:00:00.000+0000
          tipo: CONSTANCIA_SANCION
          titulo: Constancia de sanción...
          id: doc-1
          url: https://ejemplo.mx/archivo/documento.pdf
        fechaCaptura: 2017-07-21T17:32:28Z
        causaMotivoHechos: Abuso de autoridad
        observaciones: ABC
        servidorPublicoSancionado:
          puesto: Director de área
          primerApellido: Pérez
          genero:
            clave: M
            valor: MASCULINO
          segundoApellido: López
          nivel: ABCD1234
          rfc: PAT21343213
          curp: ABCD780814MDFAXL
          nombres: Carlos
        id: "1"
        inhabilitacion:
          fechaInicial: 2018-08-17T00:00:00.000+0000
          plazo: 3 meses
          fechaFinal: 2018-11-17T00:00:00.000+0000
    post_query_sort:
      type: object
      properties:
        rfc:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
        curp:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
        nombres:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
        primerApellido:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
        segundoApellido:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
        institucionDependencia:
          type: string
          description: Campo a ordenar de forma ascendente o descente
          example: asc
          enum:
          - asc
          - desc
    post_query_query:
      properties:
        id:
          type: string
          description: Devuelve resultados al buscar una declaración por su ID. Si
            se especifica este valor, los demas filtros son omitidos.
          example: 123abc
        nombres:
          type: string
          description: Nombres o nombre del servidor público sancionado.
          example: Juan Pedro
        primerApellido:
          type: string
          description: Primer apellido del servidor público sancionado.
          example: Tenorio
        segundoApellido:
          type: string
          description: Segundo apellido del servidor público sancionado.
          example: Calderon
        rfc:
          type: string
          description: RFC del Servidor Público con homoclave
          example: PAT21343213
        curp:
          type: string
          description: CURP del Servidor Público
          example: ABCD780814MDFAXL
        institucionDependencia:
          type: string
          description: Nombre de la institución dependencia a la que pertenece el
            servidor público sancionado.
          example: Secretaría de Hacienda y Crédito Público
        tipoSancion:
          uniqueItems: true
          type: array
          description: "Clave(s) del tipo de sanción aplicada. \n  * `I` - INHABILITADO\
            \ \n  * `M` - MULTADO\n  * `S` - SUSPENSIÓN DEL EMPLEO, CARGO O COMISIÓ\
            N\n  * `D` - DESTITUCIÓN DEL EMPLO, CARGO O COMISIÓN\n\n\n*******Deberá\
            \ devolver todos los registros que contengan al menos uno de los valores\
            \ solicitados.\n"
          example:
          - I
          - M
          items:
            type: string
            enum:
            - I
            - M
            - S
            - D
        rfcSolicitante:
          type: string
          description: RFC de la persona quien hace la consulta, si se omite este
            parametro, se enviaran solo los datos públicos
          example: GOAP780710RH7
    dependencias_inner:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la Institución o Dependencia donde el contratista
            cometió la irregularidad
          example: Secretaría de Hacienda y Crédito Público
        siglas:
          type: string
          description: Siglas de la Institución o Dependencia donde el contratista
            cometió la irregularidad
          example: SHCP
        clave:
          type: string
          description: Clave de la Institución o Dependencia donde el contratista
            cometió la irregularidad
          example: XYZ987
      example:
        clave: XYZ987
        siglas: SHCP
        nombre: Secretaría de Hacienda y Crédito Público
    tipoSancion_inner:
      required:
      - clave
      - valor
      type: object
      properties:
        clave:
          type: string
          description: "Tipo de sanción aplicada. \n  * `I` - INHABILITADO \n  * `M`\
            \ - MULTADO\n  * `S` - SUSPENSIÓN DEL EMPLEO, CARGO O COMISIÓN\n  * `D`\
            \ - DESTITUCIÓN DEL EMPLO, CARGO O COMISIÓN\n  * `IRSC` - INDEMNIZACIÓ\
            N RESARCITORIA\n  * `SE` - SANCIÓN ECONÓMICA\n  * `O` - OTRO\n"
          example: M
          enum:
          - I
          - M
          - S
          - D
          - IRSC
          - SE
          - O
        valor:
          type: string
          description: Valor del tipo de sanción
          example: MULTADO
        descripcion:
          type: string
          description: Descripción o nota aclaratoria
  securitySchemes:
    ssancionados_auth:
      type: oauth2
      flows:
        password:
          tokenUrl: /oauth
          scopes: {}